completed test-2
Completed part 2 of the test kata. Implemented state for the opening and closing of folders. For visual effect, I added a margin to the left of any open folders. To test whether a folder within a folder would work, I amended the data.json file to allow for a 3rd level of folder. I've also added an empty folder to see if my code works with a folder containing no other items.