Completed test-4
The only issue I had towards the end of the 4th part of the test was trying to think of a simple way to filter through all of the files, rather than those limited to the top level. I also refactored my code and cleaned up some of the variables to make them easier to understand.